test: update delete user form test
